+++ packages/react-native/android/src/main/kotlin/io/invertase/notifee/NotifeeApiModule.kt
@@ -323,6 +323,7 @@ class NotifeeApiModule(reactContext: ReactApplicationContext) :
 
     override fun addListener(eventName: String) {
         // Keep: Required for RN built in Event Emitter Calls.
+        NotifeeReactUtils.flushPendingEvents()
     }
 
     override fun removeListeners(count: Double) {